hyperliquid-funding-scraper/requirements.txt
# Core dependencies - Latest stable versions
python-dotenv==1.0.1
supabase==2.11.0
selenium==4.27.1
webdriver-manager==4.0.2

# Data processing - Python 3.13 compatible
pandas==2.2.3
numpy==2.1.2

# HTTP and networking - Latest versions
requests==2.32.3
urllib3==2.2.3
httpx==0.28.0

# Scheduling - Latest stable
schedule==1.2.2
APScheduler==3.11.0

# Logging and monitoring - Updated
coloredlogs==15.0.1
python-json-logger==2.0.7

# Utilities - Latest versions
click==8.1.7
retry==0.9.2
pytz==2024.2

# Data validation - Latest Pydantic v2
pydantic==2.10.2
pydantic-settings==2.7.0

# Testing - Latest versions
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-cov==6.0.0
faker==33.1.0

# Development tools - Latest
black==24.10.0
flake8==7.1.1
mypy==1.13.0
isort==5.13.2
pre-commit==4.0.1

# Type hints - Latest
types-requests==2.32.0.20241016
types-pytz==2024.2.0.20241003

# Additional dependencies for stability
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10